feat: implement SiteHeader and SiteHeaderClient components with dynamic category fetching and responsive design
